requirements.txt
torch>=2.1.0
transformers>=4.36.0
trl>=0.7.0
peft>=0.7.0
accelerate>=0.25.0
datasets>=2.14.0
bitsandbytes>=0.41.0
huggingface_hub
gradio>=4.0.0
sentencepiece
protobuf
python-dotenv
scipy
einops
matplotlib
